requirements.txt
streamlit
python-dotenv
open ai
